rename: UserFixture -> MemberFixture로 이름 변경
